[#183] 🚚 move : 관리자 페이지 컴포넌트 분리
